src/data/sap-ai-fabric-console/packages/api-server/requirements.txt
# SAP AI Fabric Console - API Server
# FastAPI Backend

# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# HTTP Client — async health probes and service calls
httpx>=0.26.0
aiohttp>=3.9.0

# Prometheus metrics
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0

# AI/ML — LangChain integration
langchain>=0.1.0
langchain-community>=0.0.10

# HANA Cloud Vector Engine
hdbcli>=2.19.21

# KùzuDB — graph lineage store
kuzu>=0.3.2

# Utilities
python-dotenv>=1.0.0
structlog>=24.1.0
tenacity>=8.2.3

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
